Fixup #23. Make home button works correct, Remove .idea folder from repository
