+++ .gitignore
@@ -1,8 +1,8 @@
 *.iml
 .gradle
 /local.properties
-/.idea/*
 .DS_Store
+/.idea
 /build
 /captures
 .externalNativeBuild
