+++ app/.gitignore
@@ -1,4 +1,4 @@
 /build
 /release
-/.ideag
+/.idea
 keystore.properties
