+++ app/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Tue Mar 06 00:23:13 MSK 2018
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-4.1-all.zip
