+++ app/src/main/AndroidManifest.xml
@@ -12,15 +12,19 @@
         android:fullBackupContent="@xml/backup_descriptor"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
+        android:clearTaskOnLaunch="true"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme"
         android:name=".LauncherApplication"
         tools:ignore="GoogleAppIndexingWarning">
 
-        <activity
-            android:name=".ui.activities.EntryActivity"
-            android:noHistory="true" >
+        <activity android:name=".ui.activities.LauncherActivity"
+            android:launchMode="singleTask"
+            android:stateNotNeeded="true"
+            android:noHistory="true"
+            android:clearTaskOnLaunch="true" >
+
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -30,13 +34,8 @@
             </intent-filter>
         </activity>
 
-        <activity android:name=".ui.activities.LauncherActivity"
-            android:launchMode="singleTask"
-            android:stateNotNeeded="true"
-            android:noHistory="true"
-            android:clearTaskOnLaunch="true"
-            />
         <activity android:name=".ui.activities.WelcomeActivity" />
+
         <activity android:name=".ui.activities.ProfileActivity"
             android:label="@string/app_name"
             android:parentActivityName=".ui.activities.LauncherActivity">
@@ -46,11 +45,11 @@
 
             <intent-filter>
                 <action android:name="android.intent.action.VIEW" />
-                <category android:name="android.intent.category.DEFAULT" />
                 <category android:name="android.intent.category.BROWSABLE" />
                 <data android:scheme="sychyov_md" android:host="profile_activity" />
             </intent-filter>
         </activity>
+
         <activity android:name=".ui.activities.SettingsActivity" android:parentActivityName=".ui.activities.LauncherActivity">
             <meta-data android:name="android.support.PARENT_ACTIVITY"
                 android:value="com.example.dsychyov.sychyovmd.ui.activities.LauncherActivity" />
